Added the main version
This version contains the body of the code to solve the inverse problem. At this stage, it can solve the example u = u[1] * Cos(k*x[0]) using the conjugate gradient method. When the file is run, outputs (cost, misfit, regularization, gradient's norm) will be printed to the screen. The output files are written to a folder called results under the current working directory.